[M2] refactor: 로또 숫자 생성 역할 분리
참조: https://github.com/next-step/java-lotto/pull/2521
